Add new request model. Rename response models.
